[Host IR] Allocation cache (#4466)
# What
Add an allocation cache at the HostIrEvaluator level, that is persistent
across runs. Before that, we were only relying on torch allocation
cache, but I observed that it doesn't always behave as excepted,
especially on multiprocess and multistream programs, when dealing with
large tensors.

The proposed solution may be too simple in the long run -- we should
think about how make it more controllable/smart in the future.

Illustration of the problem:
We execute a simple program with Allgather + GEMM, without even overlap.
The generated Host program contained an Allocation node for allocating
the allgather's destination buffer. Before the current patch, we observe
that, across iterations, the buffers sometimes (but not systematically)
get reallocated, thus torch's allocation cache is not hit. This is only
observed for large matrices.
![Screenshot 2025-05-16 at 18 23
37](https://github.com/user-attachments/assets/1a1cc930-f5c1-4d94-934a-95042cf2192a)