[Greedy scheduler] scatter batching support (#5274)
For #5192 

Adds support of batching for ScatterOp. Since a single ScatterOp
effectively needs two scheduling parameters, one for the index logical
domain and another for the input and the consumers of the output tensor,
`GreedyParams` is extended to store parameters of a tensor both as a
producer and a consumer.

With this PR, the greedy scheduler reduces the number of segments of the
current test_moe.py from 22 to 16, out of which 7 are linears and
grouped GEMMs and a few no-op segments. After this PR, I'll focus on
turning on the schedule by default.

test_moe.py segmentation result looks like:

<img width="1524" height="6520" alt="moe"
src="https://github.com/user-attachments/assets/ec718fc4-7424-4ef0-bb87-ec1faaf6820a"
/>