Allow allocation and loop to have different positions of device IDs (#5323)
The previous implementation ensured that allocation and loop domains
were identical due to limitations in stack (#4381). With [recent
changes](https://github.com/NVIDIA/Fuser/issues/4381#issuecomment-3309951831),
we are able to support different allocation and loop domains, and this
reordering is unnecessary.

This PR reorders device IDs to the front of the loop domain. They are at
the position of origin in allocation domain.

Following PRs will update this pass for `Stream` parallel type.

**Benchmark results: No notable difference.**

This table compares the maximum of the minimum time across all ranks for
the Transformer forward and backward passes. Results are for 8 H100
GPUs.

| Test | Main | PR |
| :--- | :---: | :---: |
| **Forward (`test_transformer_forward`)** | 2.2129 | 2.2107 |
| **Backward (`test_transformer_backward`)** | 4.1178 | 4.1033 |

---