Support layout op in scheduler (#5174)
## Stacked PRs

#5230 moe layer with nvfp4 grouped_mm
#5345 exposing layout op at direct python binding
#5198 refactor number of groups in layout op
#5174 allow layout op in automatic scheduler  <-- this PR

## This PR

Allow scheduler to take `PreprocessGroupedMatmulInputSf` as a pointwise
operation using the runtime function.

The main code change is to addressing the assumption of the runtime
function:

- [x] add segmentation for offsets to ensure they are in global memory.
* Existing assumption is that two offsets inputs and output of the
layout op would be in global memory, where the runtime function could
read the entirety of both offsets and write the output via data
dependent indexing. This allows the operation to be treated as a trivial
pointwise-op.
    * avoids caching layout op outputs or offsets inputs.
* avoids putting layout op output into persistent buffers (since we
require write to global memory).

- [x] detect unsafe consumption of PreprocessGroupedMatmulInputSf output
in `fusion_segmenter.cpp`
- [x] relax asserts on assumption that there's always a legit path
between loop->allocation and logical->allocation in some scheduler
utils.


TODOs for future PR:

* end-2-end python test with direct binding.