Fix inlining positions with constrained ops (#5317)
This was an oversight when the greedy scheduler was extended with
batching. The uninlinable IDs need to be loop IDs, whereas before this
PR they stayed at the logical domain. Generated code didn't result in
any error because the loop IDs are either parallelized with TIDx or
Group, but still limiting the inlining position to the left of
constrained IDs should make more sense.