Fix predicate elimination with initialized tensors (#5355)
In our current codegen, ArgsortOp and ScanOp inputs need to be properly
initialized so that CUB-based runtime functions can run without any
predicate. The predicate elimination needs to be aware of the
initialization and avoid omitting their predicates. For example, in the
code below, `T2` is the input of the block argsort. It's initialized to
`POS_INFINITY` as the argsort is ascending sort.

```
  Array<float, 4, 1> T2;
  #pragma unroll
  for(nvfuser_index_t i5 = 0; i5 < 4; ++i5) {
    T2[i5] = POS_INFINITY;
  }
  #pragma unroll
  for(nvfuser_index_t i5 = 0; i5 < 4; ++i5) {
    if ((i2 < (-(i5 + nvfuser_zero)))) {
      T2[i5]
         = T1[i5];
    }
  }
  Array<int64_t, 4, 1> T3;
  argsort::blockArgsort<26, 1, 1, 0, 1, 1, float, 4>(*(int64_t(*)[4])(&T3[0]), *(float(*)[4])(&T2[0]), false, DefaultBlockDim());
```

The above code is the code generated after this PR's fix. Without the
fix, the predicate of `if ((i2 < (-(i5 + nvfuser_zero)))) {` is
incorrectly omitted.

I haven't realized this issue until recently because I only thought
about cases where the producer of the argsort input is a fusion input,
i.e., `T1` of the above case. In such a case, the predicate elimination
always keeps the predicate, so this problem hasn't showed up.
Specifically, the modified test was fine previously without the
modification, but inserting a `set` exposes the issue.