Visit allIDs in IterVisitor (#5384)
This is split off from #5345, so I don't have a specific repro for any
incorrect behavior that this fixes.

Previously, we only traversed all producers of the loop domain in
IterVisitor::traverseBetween. That is a problem in cases where we
schedule like producer of a reshape, or in exotic cases like #5345 where
the domains are disconnected.

This PR ensures that we traverse every ID in the TensorDomain regardless
of the relations between the domains contained within.

---------

Co-authored-by: jjsjann123 <jiej@nvidia.com>