refactor number of groups in layout op (#5198)
## Stacked PRs

#5230 moe layer with nvfp4 grouped_mm
#5345 exposing layout op at direct python binding
#5198 refactor number of groups in layout op  <-- this PR
#5174 allow layout op in automatic scheduler

## This PR

This is a tiny refactor to only expect the two `offsets` to have the
size equal to num_groups. The reason is that our cutlass kernel were
expecting that in the first place and I didn't match it right in the
first time.

e.g. with total sequence length 10, and tokens per expert [2, 3, 5]
Previously the offsets would be [0, 2, 5, 10]; after the refactor, the
offsets would be [0, 2, 5].