Allocate CUPTI buffer dynamically (#5402)
CUPTI 13 has changed the default behavior for multi-thread execution.
From the release notes:
> By default, CUPTI will request activity buffers separately for each
thread that generates activity records. This will improve the runtime
performance of the application as it avoids contention when multiple
threads are generating activities concurrently.

From CUPTI 12.3 notes:
> Several performance improvements are done in the tracing path. One of
the key improvements is to allow clients to request CUPTI to maintain
the activity buffers at the thread level instead of global buffers. This
can be achieved by setting the option
CUPTI_ACTIVITY_ATTR_PER_THREAD_ACTIVITY_BUFFER of the enum
CUpti_ActivityAttribute. This can help in reducing the collection
overhead for applications which launch CUDA activities from multiple
host threads.

This change in default behavior means that each thread requests the
buffer, and calls `cupti_buffer_complete`. With the static buffer
allocation, this implied we read the same buffer from the beginning
multiple times, and hence records were processed again. This caused the
error where we saw the same correlation ID: Issue #4685.

Based on the recommendation of the release notes and CUPTI examples I am
changing this to dynamic allocation of buffers.