Support flexible fusion input order and workspace size in Cutlass codegen GEMM (#5387)
This does some general cleanup for the Cutlass codegen. It also enables
us to send in the fusion inputs in any order. Also, we previously
skipped initializing a workspace, assuming it was empty. Now we allocate
the workspace in the executor when running the kernel. These changes
will make it easier to support epilogue inputs in a follow-up PR.

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>