Fix and Reenable Ring Allgather Cuda Ipc Test (#5429)
PR https://github.com/NVIDIA/Fuser/pull/5325 disabled it after removing
the barrier in ipc exchangeHandles.

It was hard to reason through, but I realized that the get zcpy protocol
aligns nicely with the way the ring allgather algorithm works here. We
1) signal that the current buffer is ready to be get, and 2) matmul it
on the same stream after signaling, and 3) get the next buffer on the
next stream. On the next j iteration, the buffer is ready and we're back
at step 1.

The semantics of the get protocol allows the removal of the sendWait and
recvWait steps in the algorithm loop. I think the put protocol can do
something similar, but the algorithm would need to be rewritten so that
it's working with the current and previous buffers in the ring, instead
of current and the next buffers. For now I just skipped the test if the
put protocol is enabled.