Limit loop promotion analysis to those IDs that are in the logical-loop paths (#5451)
WAR for a loop promotion issue reported in #5391. The issue happens
because of `preprocessGroupedMatmulInputSf`, which creates an allocation
domain that is not connected with the logical domain, which is the real
cause of the issue, but to avoid it, an easy workaround is not doing
anything with allocation domains in the loop promotion analysis because
loop promotion just needs to analyze logical (root) to loop domains.

The fix here is first finding all IDs between logical and loop domains
and when only those IDs are considered propagating info. Since the
disconnected allocation ID is no longer visited, the failure as seen in
#5391 shouldn't happen.

This doesn't mean the disconnected allocation ID has no issue. We should
still make it represented more properly but it would require, for
example, tensor list support.