Support block-scaled outputs in Cutlass EVT (#5441)
This change adds support for block-scaled outputs in the Cutlass+EVT
scheduler. This is done by pattern matching for the Fusion patterns in
https://github.com/NVIDIA/Fuser/blob/94a348866f496e42a5bb2a3fd2e893ea3d2952a9/tests/cpp/test_low_precision_recipe.cpp.
This includes mxfp8 and nvfp4 outputs with and without a global scale
factor (lagged i.e. passed as an input to the Fusion).

When an output matches this pattern, we emit an
`Sm100BlockScaleFactorRowStore` EVT node. That node outputs scale
factors using a data pointer to the scale factor tensor passed as a
runtime argument. The node then passes the quantized values out. When
this node is the root of the EVT that means the regular output of the
gemm is the quantized values, having saved the scale factors along the
way as a side effect of the new node.

One complication is that the scale factor EVT node requires a device
pointer to a normalization constant. Usually this will be the numeric
limit max of the quantized dtype, e.g. 6.0 for nvfp4. However, I think
this is also the mechanism by which we could use per-tensor amax as
well. For now, I'm enabling detecting the global amax pattern and since
the EVT node requires it, I refuse to translate if global amax is not
found. In the future we could build this kind of thing into the executor
by mapping scalar values to some persistent scalar tensors on the vice
managed by the executor itself.

---------

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>