[Greedy scheduler] Make the topk support more flexible  (#5478)
The greedy scheduler processes `TopKOp` with TIDx parallelization only,
and as of ToT, all of the TIDx threads must participate in the CUB
device call with a valid input. In other words, if there are more
threads than necessary, it would be rejected. The added test would be
segmented with the current main branch due to this.

This PR relaxes the constraint by always initializing `TopK` inputs.
This follows the same approach as is done for `ArgsortOp` and `Scan`.