Add validation requirements for the block quantization op. (#5468)
Basic tests:
	
1. Input is in local memory.
Check the input TV is in local memory
2. Block scaling factor is in global memory and quantized output is in
local memory.
Check the TVs are in global/local memory
  
Schedule related tests:

1. Any loop ID that is not TIDx, TIDy, BIDx. BIDy or Group has an extent
of 1.
This is because we don’t want the device function to be called in a loop
inside the kernel since we don’t know how this will affect the indexing
inside the device function.
2. The Group ID has an extent of 4/8 depending on the data type.
This is easy to check. Get the ID corresponding to the parallel type and
check the extent.
3. There are no  TIDz/BIDz IDs.
We don’t handle 3D scheduling. Checking the parallel types should be
sufficient to test this.
4. The quantized output does not have an allocation domain.

More involved schedule related tests:

Indexing the block scaling factor output in the runtime function:

We (almost) compute the index of the block scaling factor and pass it to
the device/runtime function.
We compute the linearized index in index.cpp. We get the logical indices
for quantized output and then compute the linearized indices with the
extents of the logical domain.

```
std::vector<Val*> logical_index = Index::getConsumerPerDimLogicalIndex(
      bqop->quantizedOutput()->as<TensorView>(), for_loops_, getRotatedLoop());

  auto loop_domain =
      bqop->quantizedOutput()->as<TensorView>()->getLogicalDomain();

  int64_t dim_count = logical_index.size();

  // logical_index[2] * 1 + logical_index[1] * extent[2] + logical_index[0] *
  // extent[1] * extent[2]
  auto idx = logical_index[dim_count - 1];
  for (auto i = dim_count - 2; i >= 0; i--) {
    auto stride = IrBuilder::create<Val>(1, DataType::Index);
    for (auto j = i + 1; j < dim_count; j++) {
      stride = IrBuilder::mulExpr(stride, loop_domain[j]->extent());
    }
    idx = IrBuilder::addExpr(IrBuilder::mulExpr(logical_index[i], stride), idx);
  }
```

We pass the index to the runtime function, where we divide this by 16
(blocksize) to index the block scaling tensorview..

To make the above indexing scheme work we need to place a couple
restrictions on the schedule of Block Quantization which we need to
validate. They are:

1. The Group ID should be the fastest changing dimensions. 
2. The next ID should be TIDx
3. All merges from logical domain to Group ID and Thread ID should be
contiguous merges.

We illustrate with an example of a valid schedule:

<img width="815" height="712" alt="Screenshot 2025-11-03 at 9 40 52 AM"
src="https://github.com/user-attachments/assets/b23f8594-dfba-4414-a3e5-4479732765aa"
/>

Implementation:
We find the transforms from the logical domain to the loop domain of the
quantized output. We apply these transforms to the logical domain to get
a loop domain (minus any reordering). While applying these transforms
(merges and split) we track all IDs produced without non-contiguous
merge. In this loop-like domain produced after applying transforms, we
make sure that Group ID is inner-most and Thread ID (x) is the next
inner-most. If there are other IDs between them, their extent should be
1.

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>