Extend the block quantization runtime function to also handle 2 and 8 (bf16/pf16) elements per thread (#5488)
Currently the runtime function that implements block quantization to fp4
is limited to working on 4 elements per thread.
This change enables it to work on 2, 4 and 8(bfloat) elements per
thread.

Most of the updates were to the runtime function, where the code using
the shuffle instruction were updated.
The other small changes are relax the restriction on the number of
elements per thread and enable codegen to call the runtime function
correctly.

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>