Do not parallelize new iterdomains on device (#5497)
While working on `DecomposeReshardingsPass`, the new tensorviews have
ids that are parallelized through the ops API. This caused the following
issue:
1. When using the logic from `PropagateShardingsPass` (PR #5496), the
already parallelized IDs are not overwritten.
2. The alternative is to unparallelize in `DecomposeReshardingsPass`.
However, I think automatically parallelizing iterdomains in the ops API
is unnecessary. It also causes a divergence in loop vs logical split, as
we don't propagate the loop split in the ops API. Replaying
parallelization and transforms can be confined to utils.