Clone all expressions before lowering for simplicity (#5512)
This has two benefits: 
1. avoid the recomputeTv hack
2. make it easier to inline multiple segments into the same for-loop:
https://github.com/NVIDIA/Fuser/pull/5514. We'll need to create
for-loops **outside** lowerSegment with **cloned** IterDomains. It's too
late to clone **inside** lowerSegment.

The downside is that we are cloning more instructions than necessary
into the HostIrContainer. For example, without this PR, expressions in a
kernel segment are not cloned -- they are replaced with a LaunchKernel
instruction.

Overall, I think it's a net win -- I prefer simplicity unless the
performance benefit is well-demonstrated.

For #5229