Refactor Cutlass codegen to support compile time check. (#5494)
This is motivated by needing an accurate compile time check for the
Cutlass scheduler. The conditions for accepting a fusion are complicated
and evolve whenever we support new features in EVT. Instead of trying to
keep a separate `canScheduleCompileTime` up to date, this change throws
a special exception whenever we hit an unsupported pattern, allowing us
to guarantee that we will reject any fusion we can't generate code for.
This means that we will get relevant messages like this in segmentation
when we can't accept a segment:
```
Scheduler _cutlass_ ***rejected*** because : Bias not yet supported for EVT translation
```

Specific changes:
1. Introduce `UnsupportedFusion` exception type and
`NVF_CUTLASS_REJECT{,_IF}` macros that will check conditions and throw
that exception type analogous to `NVF_ERROR` and `NVF_THROW`.
2. Update code in `csrc/cutlass/*` to use the new macros instead of
`NVF_ERROR` where appropriate.
3. Refactor `csrc/cutlass/gemm.cpp` to use a helper class. This cleanup
was overdue anyway, but it makes it a bit clearer to operate in two
stages: one where we gather fusion info and another where we generate
the actual code.
4. Update the cutlass scheduler's `canScheduleCompileTime` to run just
the info gathering stage of codegen in order to check for exceptions.
5. Make a few things static in the generated kernels: no longer template
`args_from_inputs` and statically compute alignment in bits in nvFuser
instead of doing it using cutlass.

No practical change to generated code is expected from this change.