Use pointwise scheduler to schedule Block Quantization (#5362)
Stacked on top of: https://github.com/NVIDIA/Fuser/pull/5488

Goal: To enable the pointwise scheduler to accept the block quantization
operator

To handle the block quantization operator I have made the following
assumptions/changes which I would like feedback on.

1. If there's a block quantization operator, do not unroll
2. When trying to find a reference tensor to schedule, ignore the block
scaling factor output of the block quantization op as it has unmapped
IDs.
3. We check that the block scaling factor is an output of the fusion
otherwise the pointwise scheduler will reject.
4. If we don't find that find that we could vectorize the fusion segment
by a factor of 2 or more, then we throw.
5. We first vectorize the quantized output. Later the vectorized ID is
changed to a Group Id. May be there should be a better way to directly
mark that ID as Group.

Implementation:

When scheduling, we let scheduling process as usual (with the changes
listed above).
For vectorization, we first vectorize the quantized output of the block
quantization op, and then change the vectorized ID to a group ID.

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>