Enable TensorIndexer with the transpose tests (#4246)
Enabling TensorIndexer for the transpose tests. Diff check revealed some
mismatches, but I believe they are benign. As far as I can see, all of
them are caused by some slight difference for issue #681.
TensorIndexer's workaround is simpler but not as precise as the legacy
indexer, as indicated in #2689.

For example, there is an index for a predicate of
`TransposeTest.FusionScheduleTransposeSimple`:

```
 ( ( ( blockIdx.x % ( ( ceilDiv(( (( (( getMetaData(T0) )).logical_size ))[2] ), 32) ) * ( ceilDiv(( (( (( getMetaData(T0) )).logical_size ))[1] ), 32) ) ) ) / ( ceilDiv(( (( (( getMetaData(T0) )).logical_size ))[1] ), 32) ) ) * 32 ) + (threadIdx.x * 4) % 32)
```

The above is generated by the legacy indexer. With TensorIndexer, it's
generated as:

```
( ( ( ( blockIdx.x % ( ( ceilDiv(( (( (( getMetaData(T0) )).logical_size ))[2] ), 32) ) * ( ceilDiv(( (( (( getMetaData(T0) )).logical_size ))[1] ), 32) ) ) ) / ( ceilDiv(( (( (( getMetaData(T0) )).logical_size ))[1] ), 32) ) ) * 32 ) + 0 )
```

The difference is `threadIdx.x 4 % 32` vs. `0`. Note that this is a
predicate for the condition at the start position, i.e., they are
predicated with `>= 0`, and as such the index of TensorIndexer is more
conservative, so the legacy indexer may have a higher chance of using
the fast path created by the unswitch predicate. However, in practice,
both predicates should be always true, so it shouldn't matter. I don't
see any actual difference with stop predicates.