Pull the Llama4 inference benchmark from lightning-thunder (#5578)
Pull the latest version of this script after substantial changes in the
lightning repo. The script is being "moved" into the fuser repo - it
will be deleted from lightning-thunder, and subsequent changes will be
merged into the fuser repo.

The script does not have any nvfuser changes at this moment. No configs
or model code are pulled inside yet.