Simplify argument passing in Cutlass+EVT executor (#5577)
This is strictly a cleanup PR and does not change any functionality in
generated kernels.

There are two changes here: a small cleanup to function signatures in
CutlassCompiledKernel, and a larger change to `csrc/cutlass/gemm.cpp`
which affects code generation. The main change there is to simplify and
clarify argument names in the generated kernel using an `Inputs` struct
which is passed to the relevant functions. The kernel holds m, n, k, and
device pointers to the relevant gmem buffers. This is similar
information to what is held in the cutlass arguments data structure, but
it is a little easier for us to manipulate.

For grouped GEMM (see #5528) we will also use this `Inputs` struct to
pass arguments to the `get_group_gemm_starts` kernel.

---------

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>