fix nvfp4 in reference implementation in llama4 benchmark (#5642)
A few things patched:

1. offsets was updated in old implementation to only contain number of
group size, so we shouldn't slice it again in forward method;
2. quantization is done assume fast dimension is on the innermost.
Needed to transpose fp4 tensor before and after the quantization;
3. switched dequantization implementation to TE for much faster
performance (the old implementation hangs for many minutes).
4. minor fix on import path.

With these fixes, at least inference with nvfp4 via thunder/eager is
functional. Though performance is much slower, because we don't have a
function nvfp4 grouped_mm kernel and we have to dequantize the weight in
order to run bf16 grouped_mm.

---------

Co-authored-by: Protonu <pbasu@nvidia.com>