Fix detection of ldmatrix/stmatrix (#5645)
This was found while I was experimenting TensorIndexer with the matmul
tests (#5574). Ldmatrix and stmatrix use a special domain as an
alternative loop domain for indexing. IIUC, we should not use the
alternate domains when initializing tensors. This happens, for example,
a tensor is defined by an stmatrix op but is also initialized to zero
for predicate elimination. Looks like the initialization should not be
done at all, but I think that's a separate issue.

Please see https://github.com/NVIDIA/Fuser/pull/5645/files#r2600720284.
The other changes are just due to this change.

---------

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>