Enable TensorIndexer for the matmul tests (#5574)
Enabling TensorIndexer with all matmul tests. The codegen diff shows
quite many tests exhibiting some differences. While I haven't examined
everything, looks like the changes come from the difference of the
initialization of buffers. For example, in
`MmaTest/HopperRSStmatrix.SingleTileWithTMALoadStoreStMatrix/4`, here's
the diff result:

```
@@ -12793,11 +12793,11 @@

   nvfuser_index_t i6;
   i6 = 2 * (((nvfuser_index_t)threadIdx.x) % 4);
   nvfuser_index_t i7;
   i7 = ((256 * i4) + (16 * i5)) + i6;
   nvfuser_index_t i8;
-  i8 = ((384 * i4) + (24 * i5)) + i6;
+  i8 = 4 * ((nvfuser_index_t)threadIdx.x);
   bool b9;
   b9 = ((nvfuser_index_t)threadIdx.x) < 32ULL;
   nvfuser_index_t i10;
   i10 = 16 * i4;
   nvfuser_index_t i11;
@@ -12897,11 +12897,11 @@

       }
     }
   }
   #pragma unroll
   for(nvfuser_index_t i39 = 0; i39 < 3; ++i39) {
-    arraySet<__half, 4>(&T5[(i8 + (8 * i39))], (__half)0);
+    arraySet<__half, 4>(&T5[(i8 + (512 * i39))], (__half)0);
   }
   #pragma unroll
   for(nvfuser_index_t i39 = 0; i39 < 3; ++i39) {
     if ((b13 && (i14 < (-(8 * i39))))) {
       stmatrix2((uint32_t)((toSmem(T5) + ((48 * (((nvfuser_index_t)threadIdx.x) % 32)) + ((((nvfuser_index_t)threadIdx.y) * 3072LL) + ((((i39 % 1) * 3072) + ((i39 / 1) * 16)) + ((((((nvfuser_index_t)threadIdx.x) / 32) % 4) * 768) + (((((nvfuser_index_t)threadIdx.x) / 32) / 4) * 16))))))), (*reinterpret_cast<Array<uint32_t, 2, 1>*>(&T7[(4 * i39)])));
```

We probably should not need to do this initialization (see #5657), but
for the sake of this PR, the change here should be benign. Previously,
our indexing is always based on the domain returned by
`getMaybeAllocationDomain`, whereas TensorIndexer uses the actual
allocation, which can be the loop domain, so this difference of
generated code is expected.