Extend the block quantization op to quantize to mMXFP8. (#5643)
We can use the block quantization op to quantize FP32/BF16/FP16 to
MXFP8.

1. We implement the quantization with the runtime function in
block_quantization_kernels.cu
2. In our Python tests we get bitwise parity with TE.
3. This PR modifies our __float2e8m0 cast to round up (`rp`) to match
TE.

There will be a follow-up PR to modify the scheduling for Block
Quantization op in the case of quantizing to MXFP8.
So far we need vectorization (of >=2) for block quantization. We should
be able to support quantization to MXFP8 without the need for any
vectorization.

---------

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>